+++ docs/CHANGE_LOG.md
@@ -2,8 +2,11 @@
 
 ## Version 2.2.0:
 
+#### **Public Functions:**
+- Added "includeDataAttributes()" support for "json()" (including binding support), which states if the data attributes should be included (defaults to true).
+
 #### **Documentation:**
-- Updated the documentation to show the right properties for the objects.
+- Updated the documentation to show the right properties for the objects, and add missing documentation for the binding attribute.
 
 <br>